Introduced rules to make Makefile lighter; small restructuring
- Added rules_*.mk containing source and header files
- Moved boards/megafly_rev4.* and proj_avr32/config to boards/megafly_rev4/
- Removed unused files of proj_avr32/config
