+++ boards/megafly_rev4/conf_sleepmgr.h
