+++ boards/megafly_rev4/conf_usb.h
