+++ boards/megafly_rev4/megafly_rev4.cpp
@@ -38,7 +38,7 @@
  *
  ******************************************************************************/
 
-#include "boards/megafly_rev4.hpp"
+#include "boards/megafly_rev4/megafly_rev4.hpp"
 
 extern "C"
 {
