+++ rules_avr32.mk
@@ -1,3 +1,34 @@
+LIB_SRCS += boards/megafly_rev4/megafly_rev4.cpp
+
+LIB_SRCS += hal/avr32/adc_avr32.cpp          
+LIB_SRCS += hal/avr32/adc_int.c          
+LIB_SRCS += hal/avr32/analog_monitor.c            
+LIB_SRCS += hal/avr32/analog_monitor_telemetry.cpp                  
+LIB_SRCS += hal/avr32/dac_dma.c
+LIB_SRCS += hal/avr32/i2c_avr32.cpp            
+LIB_SRCS += hal/avr32/led.c            
+LIB_SRCS += hal/avr32/led_avr32.cpp            
+LIB_SRCS += hal/avr32/gpio_avr32.cpp
+LIB_SRCS += hal/avr32/piezo_speaker.c        
+LIB_SRCS += hal/avr32/pwm_avr32.cpp
+LIB_SRCS += hal/avr32/serial_avr32.cpp
+LIB_SRCS += hal/avr32/serial_usb_avr32.cpp
+LIB_SRCS += hal/avr32/spi_avr32.cmegapp
+LIB_SRCS += hal/avr32/sd_spi.c
+LIB_SRCS += hal/avr32/time_keeper.c
+LIB_SRCS += hal/avr32/file_flash_avr32.cpp
+LIB_SRCS += hal/avr32/file_fat_fs.cpp
+LIB_SRCS += hal/avr32/fat_fs_mounting.cpp
+LIB_SRCS += hal/avr32/FatFs/atadrive.c
+LIB_SRCS += hal/avr32/FatFs/sdcard.c
+LIB_SRCS += hal/avr32/FatFs/usbdisk.c
+
+LIB_MEGAFLY_INC += boards/megafly_rev4
+
+# ------------------------------------------------------------------------------
+# ASF Library
+# ------------------------------------------------------------------------------
+
 LIB_ASF_INC += libs/asf
 LIB_ASF_INC += libs/asf/avr32
 LIB_ASF_INC += libs/asf/avr32/services
@@ -106,38 +137,38 @@ LIB_ASF_INC += libs/asf/common/utils/interrupt
 LIB_ASF_INC += libs/asf/common/boards
 LIB_ASF_INC += libs/asf/common/boards/user_board
 
-LIB_ASF_SRCS += libs/asf/avr32/drivers/intc/exception_noNMI.S
-LIB_ASF_SRCS += libs/asf/avr32/utils/startup/startup_uc3.S
-LIB_ASF_SRCS += libs/asf/avr32/utils/startup/trampoline_uc3.S
-LIB_ASF_SRCS += libs/asf/avr32/drivers/adcifa/adcifa.c
-LIB_ASF_SRCS += libs/asf/avr32/drivers/ast/ast.c
-LIB_ASF_SRCS += libs/asf/avr32/drivers/dacifb/dacifb.c
-LIB_ASF_SRCS += libs/asf/avr32/drivers/eic/eic.c
-LIB_ASF_SRCS += libs/asf/avr32/drivers/flashc/flashc.c
-LIB_ASF_SRCS += libs/asf/avr32/drivers/gpio/gpio.c
-LIB_ASF_SRCS += libs/asf/avr32/drivers/intc/intc.c
-LIB_ASF_SRCS += libs/asf/avr32/drivers/pdca/pdca.c
-LIB_ASF_SRCS += libs/asf/avr32/drivers/pevc/pevc.c
-LIB_ASF_SRCS += libs/asf/avr32/drivers/pm/pm_uc3c.c  
-LIB_ASF_SRCS += libs/asf/avr32/drivers/pm/power_clocks_lib.c
-LIB_ASF_SRCS += libs/asf/avr32/drivers/pwm/pwm4.c
-LIB_ASF_SRCS += libs/asf/avr32/drivers/scif/scif_uc3c.c
-LIB_ASF_SRCS += libs/asf/avr32/drivers/spi/spi.c
-LIB_ASF_SRCS += libs/asf/avr32/drivers/tc/tc.c
-LIB_ASF_SRCS += libs/asf/avr32/drivers/twim/twim.c
-LIB_ASF_SRCS += libs/asf/avr32/drivers/usart/usart.c
-LIB_ASF_SRCS += libs/asf/avr32/drivers/usbc/usbc_device.c
-LIB_ASF_SRCS += libs/asf/avr32/services/delay/delay.c
-LIB_ASF_SRCS += libs/asf/avr32/utils/startup/startup_uc3.S
-LIB_ASF_SRCS += libs/asf/common/boards/user_board/init.c
-LIB_ASF_SRCS += libs/asf/common/services/clock/uc3c/osc.c
-LIB_ASF_SRCS += libs/asf/common/services/clock/uc3c/pll.c
-LIB_ASF_SRCS += libs/asf/common/services/clock/uc3c/sysclk.c
-LIB_ASF_SRCS += libs/asf/common/services/sleepmgr/uc3/sleepmgr.c
-LIB_ASF_SRCS += libs/asf/common/services/spi/uc3_spi/spi_master.c
-LIB_ASF_SRCS += libs/asf/common/services/usb/class/cdc/device/udi_cdc.c
-LIB_ASF_SRCS += libs/asf/common/services/usb/class/cdc/device/udi_cdc_desc.c
-LIB_ASF_SRCS += libs/asf/common/services/usb/udc/udc.c
-LIB_ASF_SRCS += libs/asf/common/utils/stdio/read.c
-LIB_ASF_SRCS += libs/asf/common/utils/stdio/write.c
-LIB_ASF_SRCS += libs/asf/common/utils/stdio/stdio_usb/stdio_usb.c
+LIB_SRCS += libs/asf/avr32/drivers/intc/exception_noNMI.S
+LIB_SRCS += libs/asf/avr32/utils/startup/startup_uc3.S
+LIB_SRCS += libs/asf/avr32/utils/startup/trampoline_uc3.S
+LIB_SRCS += libs/asf/avr32/drivers/adcifa/adcifa.c
+LIB_SRCS += libs/asf/avr32/drivers/ast/ast.c
+LIB_SRCS += libs/asf/avr32/drivers/dacifb/dacifb.c
+LIB_SRCS += libs/asf/avr32/drivers/eic/eic.c
+LIB_SRCS += libs/asf/avr32/drivers/flashc/flashc.c
+LIB_SRCS += libs/asf/avr32/drivers/gpio/gpio.c
+LIB_SRCS += libs/asf/avr32/drivers/intc/intc.c
+LIB_SRCS += libs/asf/avr32/drivers/pdca/pdca.c
+LIB_SRCS += libs/asf/avr32/drivers/pevc/pevc.c
+LIB_SRCS += libs/asf/avr32/drivers/pm/pm_uc3c.c  
+LIB_SRCS += libs/asf/avr32/drivers/pm/power_clocks_lib.c
+LIB_SRCS += libs/asf/avr32/drivers/pwm/pwm4.c
+LIB_SRCS += libs/asf/avr32/drivers/scif/scif_uc3c.c
+LIB_SRCS += libs/asf/avr32/drivers/spi/spi.c
+LIB_SRCS += libs/asf/avr32/drivers/tc/tc.c
+LIB_SRCS += libs/asf/avr32/drivers/twim/twim.c
+LIB_SRCS += libs/asf/avr32/drivers/usart/usart.c
+LIB_SRCS += libs/asf/avr32/drivers/usbc/usbc_device.c
+LIB_SRCS += libs/asf/avr32/services/delay/delay.c
+LIB_SRCS += libs/asf/avr32/utils/startup/startup_uc3.S
+LIB_SRCS += libs/asf/common/boards/user_board/init.c
+LIB_SRCS += libs/asf/common/services/clock/uc3c/osc.c
+LIB_SRCS += libs/asf/common/services/clock/uc3c/pll.c
+LIB_SRCS += libs/asf/common/services/clock/uc3c/sysclk.c
+LIB_SRCS += libs/asf/common/services/sleepmgr/uc3/sleepmgr.c
+LIB_SRCS += libs/asf/common/services/spi/uc3_spi/spi_master.c
+LIB_SRCS += libs/asf/common/services/usb/class/cdc/device/udi_cdc.c
+LIB_SRCS += libs/asf/common/services/usb/class/cdc/device/udi_cdc_desc.c
+LIB_SRCS += libs/asf/common/services/usb/udc/udc.c
+LIB_SRCS += libs/asf/common/utils/stdio/read.c
+LIB_SRCS += libs/asf/common/utils/stdio/write.c
+LIB_SRCS += libs/asf/common/utils/stdio/stdio_usb/stdio_usb.c
