+++ rules_dummy.mk
@@ -1,0 +1,7 @@
+LIB_SRCS += hal/dummy/FatFs/atadrive.c
+LIB_SRCS += hal/dummy/FatFs/sdcard.c
+LIB_SRCS += hal/dummy/FatFs/usbdisk.c
+
+LIB_SRCS += hal/dummy/adc_dummy.cpp
+LIB_SRCS += hal/dummy/i2c_dummy.cpp
+LIB_SRCS += hal/dummy/gpio_dummy.cpp
