+++ rules_libs.mk
@@ -1,0 +1,5 @@
+LIB_SRCS += libs/FatFs/src/ff.c  
+LIB_SRCS += libs/FatFs/src/diskio.c
+LIB_SRCS += libs/FatFs/src/option/syscall.c
+LIB_SRCS += libs/FatFs/src/option/unicode.c
+LIB_SRCS += libs/FatFs/src/option/ccsbcs.c
