+++ rules_stm32.mk
@@ -1,0 +1,5 @@
+LIB_SRCS += boards/mavrimini.cpp
+
+LIB_SRCS += hal/stm32/time_keeper.c
+LIB_SRCS += hal/stm32/gpio_stm32.cpp
+LIB_SRCS += hal/stm32/serial_stm32.cpp
