+++ sample_projects/LEQuad/main_avr32.cpp
@@ -39,7 +39,7 @@
  ******************************************************************************/
  
 #include "sample_projects/LEQuad/central_data.hpp"
-#include "boards/megafly_rev4.hpp"
+#include "boards/megafly_rev4/megafly_rev4.hpp"
 #include "sample_projects/LEQuad/mavlink_telemetry.hpp"
 #include "sample_projects/LEQuad/tasks.hpp"
