--- sample_projects/LEQuad/proj_avr32/config/conf_constants.h
@@ -1,21 +1,0 @@
-
-/*
- * conf_constants.h
- *
- * Created: 10/23/2014 5:58:53 PM
- *  Author: lis
- */ 
-
-#ifdef __cplusplus
-extern "C"
-{
-#endif
-
-#define HOME_LATITUDE 46.5f
-#define HOME_LONGITUDE 6.6f
-#define HOME_ALTITUDE 400.0f
-#define GRAVITY	9.8f
-
-#ifdef __cplusplus
-}
-#endif
